# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Transit-AI Backend",
    "description": "Прогноз пассажиропотока трамваев Москвы (hackathon Transit-AI).",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/predictions/regenerate": {
      "post": {
        "tags": [
          "predictions-runs"
        ],
        "summary": "T-230: сгенерировать новый набор",
        "description": "Запускает Celery-генерацию прогнозов с параметрами из UI/БД.",
        "operationId": "regenerate_api_v1_predictions_regenerate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Regenerate Api V1 Predictions Regenerate Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/predictions/runs": {
      "get": {
        "tags": [
          "predictions-runs"
        ],
        "summary": "T-230: список запусков генерации",
        "description": "Последние запуски (новые сверху) + какой submission_id активен.",
        "operationId": "list_runs_api_v1_predictions_runs_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictionRunListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/predictions/runs/{run_id}": {
      "get": {
        "tags": [
          "predictions-runs"
        ],
        "summary": "T-230: статус запуска + кандидат (при Celery SUCCESS)",
        "description": "Статус запуска. Пока Celery работает — опрашиваем; при SUCCESS ищем CSV.",
        "operationId": "get_run_api_v1_predictions_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictionRunOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/predictions/runs/{run_id}/ingest": {
      "post": {
        "tags": [
          "predictions-runs"
        ],
        "summary": "T-230: загрузить CSV кандидата в БД (и опционально активировать)",
        "description": "INSERT кандидата в `predictions`. Идемпотентность: повторный ingest → 409.",
        "operationId": "ingest_run_api_v1_predictions_runs__run_id__ingest_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Run Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunActionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/predictions/runs/{run_id}/reject": {
      "post": {
        "tags": [
          "predictions-runs"
        ],
        "summary": "T-230: отклонить кандидата (оставить эталон/текущий набор)",
        "description": "Помечает кандидата `rejected`. Активный набор НЕ меняется.",
        "operationId": "reject_run_api_v1_predictions_runs__run_id__reject_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunActionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IngestBody": {
        "properties": {
          "activate": {
            "type": "boolean",
            "title": "Activate",
            "default": true
          }
        },
        "type": "object",
        "title": "IngestBody",
        "description": "Тело POST /predictions/runs/{id}/ingest."
      },
...
      "PredictionRunListResponse": {
        "properties": {
          "runs": {
            "items": {
              "$ref": "#/components/schemas/PredictionRunOut"
            },
            "type": "array",
            "title": "Runs"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "active_submission_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Submission Id"
          }
        },
        "type": "object",
        "required": [
          "runs",
          "count"
        ],
        "title": "PredictionRunListResponse",
        "description": "Список запусков + текущий активный набор."
      },
      "PredictionRunOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "celery_task_id": {
            "type": "string",
            "title": "Celery Task Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "submission_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submission Id"
          },
          "model_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Id"
          },
          "feature_set": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feature Set"
          },
          "pipeline_kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pipeline Kind"
          },
          "row_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Row Count"
          },
          "holdout_wape_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Holdout Wape Score"
          },
          "recommendation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recommendation"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "is_etalon": {
            "type": "boolean",
            "title": "Is Etalon",
            "default": false
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": false
          },
          "csv_filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Csv Filename"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "activated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Activated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "celery_task_id",
          "status",
          "started_at"
        ],
        "title": "PredictionRunOut",
        "description": "Один запуск генерации (candidate lifecycle)."
      },
...
      "RegenerateRequest": {
        "properties": {
          "coef_weather": {
            "type": "number",
            "maximum": 3.0,
            "minimum": 0.0,
            "title": "Coef Weather",
            "default": 1.0
          },
          "coef_event": {
            "type": "number",
            "maximum": 3.0,
            "minimum": 0.0,
            "title": "Coef Event",
            "default": 1.0
          },
          "coef_season": {
            "type": "number",
            "maximum": 3.0,
            "minimum": 0.0,
            "title": "Coef Season",
            "default": 1.0
          },
          "start_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Date",
            "description": "YYYY-MM-DD"
          },
          "end_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date",
            "description": "YYYY-MM-DD"
          },
          "model_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Id"
          },
          "model_kind": {
            "type": "string",
            "pattern": "^(route_baseline|xgboost_route)$",
            "title": "Model Kind",
            "default": "xgboost_route"
          },
          "feature_set": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feature Set",
            "description": "Если None — выводится из feature_toggles"
          },
          "zeros_applied": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Zeros Applied",
            "description": "Если None — выводится из zero_overrides"
          }
        },
        "type": "object",
        "title": "RegenerateRequest",
        "description": "Параметры генерации нового набора прогнозов.\n\nВсё опционально: не заданные поля берутся из БД (feature_toggles /\nzero_overrides) и из `settings` (submission period)."
      },
...
      "RunActionResponse": {
        "properties": {
          "run": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PredictionRunOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "rows": {
            "type": "integer",
            "title": "Rows",
            "default": 0
          },
          "active_submission_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Submission Id"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": ""
          }
        },
        "type": "object",
        "title": "RunActionResponse",
        "description": "Ответ на ingest/reject/restore."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}